[PRM-795] Updated routing for Upload list page (#185)

+++ app/controllers/OtherReasonController.scala
@@ -17,13 +17,17 @@
 package controllers
 
 import config.featureSwitches.{FeatureSwitching, NonJSRouting}
+import java.time.LocalDate
 import config.{AppConfig, ErrorHandler}
 import controllers.predicates.{AuthPredicate, DataRequiredAction}
 import forms.WhenDidBecomeUnableForm
 import forms.WhyReturnSubmittedLateForm.whyReturnSubmittedLateForm
-import forms.upscan.{RemoveFileForm, UploadDocumentForm, YouHaveUploadedFilesForm}
+
+import javax.inject.Inject
+import models.upload.UploadStatusEnum.READY
+import forms.upscan.{RemoveFileForm, UploadDocumentForm, UploadListForm}
 import helpers.{FormProviderHelper, UpscanMessageHelper}
-import models.pages.{EvidencePage, WhenDidBecomeUnablePage, WhyWasReturnSubmittedLatePage, YouHaveUploadedFilesPage}
+import models.pages._
 import models.{CheckMode, Mode, NormalMode}
 import navigation.Navigation
 import play.api.data.{Form, FormError}
@@ -36,13 +40,10 @@ import uk.gov.hmrc.govukfrontend.views.viewmodels.radios.RadioItem
 import uk.gov.hmrc.play.bootstrap.frontend.controller.FrontendController
 import utils.Logger.logger
 import utils.SessionKeys
-import views.html.components.upload.YouHaveUploadedFilesPage
 import views.html.reasonableExcuseJourneys.other._
-import views.html.reasonableExcuseJourneys.other.noJs.{UploadAnotherDocumentPage, UploadFirstDocumentPage, UploadTakingLongerThanExpectedPage}
+import views.html.reasonableExcuseJourneys.other.noJs._
 import viewtils.RadioOptionHelper
 
-import java.time.LocalDate
-import javax.inject.Inject
 import scala.concurrent.{ExecutionContext, Future}
 
 class OtherReasonController @Inject()(whenDidBecomeUnablePage: WhenDidBecomeUnablePage,
@@ -51,7 +52,7 @@ class OtherReasonController @Inject()(whenDidBecomeUnablePage: WhenDidBecomeUnab
                                       uploadFirstDocumentPage: UploadFirstDocumentPage,
                                       uploadTakingLongerThanExpectedPage: UploadTakingLongerThanExpectedPage,
                                       uploadAnotherDocumentPage: UploadAnotherDocumentPage,
-                                      youHaveUploadedFilesPage: YouHaveUploadedFilesPage,
+                                      uploadListPage: UploadListPage,
                                       navigation: Navigation,
                                       upscanService: UpscanService,
                                       featureSwitching: FeatureSwitching,
@@ -109,7 +110,7 @@ class OtherReasonController @Inject()(whenDidBecomeUnablePage: WhenDidBecomeUnab
         whyReturnSubmittedLateReason => {
           logger.debug(s"[OtherReasonController][onSubmitForWhenDidBecomeUnable]" +
             s" - Adding '$whyReturnSubmittedLateReason' to session under key: ${SessionKeys.whyReturnSubmittedLate}")
-            Redirect(navigation.nextPage(WhyWasReturnSubmittedLatePage, mode, Some(whyReturnSubmittedLateReason)))
+          Redirect(navigation.nextPage(WhyWasReturnSubmittedLatePage, mode, Some(whyReturnSubmittedLateReason)))
             .addingToSession(SessionKeys.whyReturnSubmittedLate -> whyReturnSubmittedLateReason)
         })
     }
@@ -117,7 +118,7 @@ class OtherReasonController @Inject()(whenDidBecomeUnablePage: WhenDidBecomeUnab
 
   def onPageLoadForUploadEvidence(mode: Mode): Action[AnyContent] = (authorise andThen dataRequired).async {
     implicit request => {
-      if(request.cookies.get("jsenabled").isEmpty || featureSwitching.isEnabled(NonJSRouting)) {
+      if (request.cookies.get("jsenabled").isEmpty || featureSwitching.isEnabled(NonJSRouting)) {
         mode match {
           case NormalMode => Future(Redirect(controllers.routes.OtherReasonController.onPageLoadForFirstFileUpload(mode)))
           case CheckMode => Future(Redirect(controllers.routes.OtherReasonController.removeFileUpload(mode)))
@@ -126,12 +127,23 @@ class OtherReasonController @Inject()(whenDidBecomeUnablePage: WhenDidBecomeUnab
         for {
           previousUploadsState <- uploadJourneyRepository.getUploadsForJourney(request.session.get(SessionKeys.journeyId))
         } yield {
-          val previousUploads = previousUploadsState.fold("[]")(previousUploads => Json.stringify(Json.toJson(previousUploads)))
-          val initiateNextUploadUrl = controllers.routes.UpscanController.initiateCallToUpscan(request.session.get(SessionKeys.journeyId).get)
-          val getStatusUrl = controllers.routes.UpscanController.getStatusOfFileUpload(request.session.get(SessionKeys.journeyId).get, _)
-          val removeFileUrl = controllers.routes.UpscanController.removeFile(request.session.get(SessionKeys.journeyId).get, _)
-          val postAction = navigation.nextPage(EvidencePage, mode)
-          Ok(uploadEvidencePage(postAction, initiateNextUploadUrl, getStatusUrl, removeFileUrl, previousUploads))
+          if (request.cookies.get("jsenabled").isEmpty) {
+            mode match {
+              case NormalMode => Redirect(controllers.routes.OtherReasonController.onPageLoadForFirstFileUpload(mode))
+              case CheckMode => if (previousUploadsState.exists(_.count(_.fileStatus == READY) > 0)) {
+                Redirect(controllers.routes.OtherReasonController.onPageLoadForUploadComplete(mode))
+              } else {
+                Redirect(controllers.routes.OtherReasonController.onPageLoadForFirstFileUpload(mode))
+              }
+            }
+          } else {
+            val previousUploads = previousUploadsState.fold("[]")(previousUploads => Json.stringify(Json.toJson(previousUploads)))
+            val initiateNextUploadUrl = controllers.routes.UpscanController.initiateCallToUpscan(request.session.get(SessionKeys.journeyId).get)
+            val getStatusUrl = controllers.routes.UpscanController.getStatusOfFileUpload(request.session.get(SessionKeys.journeyId).get, _)
+            val removeFileUrl = controllers.routes.UpscanController.removeFile(request.session.get(SessionKeys.journeyId).get, _)
+            val postAction = navigation.nextPage(EvidencePage, mode)
+            Ok(uploadEvidencePage(postAction, initiateNextUploadUrl, getStatusUrl, removeFileUrl, previousUploads))
+          }
         }
       }
     }
@@ -139,7 +151,7 @@ class OtherReasonController @Inject()(whenDidBecomeUnablePage: WhenDidBecomeUnab
 
   def onPageLoadForFirstFileUpload(mode: Mode): Action[AnyContent] = (authorise andThen dataRequired).async {
     implicit request => {
-      if(request.cookies.get("jsenabled").isDefined && !featureSwitching.isEnabled(NonJSRouting)) {
+      if (request.cookies.get("jsenabled").isDefined && !featureSwitching.isEnabled(NonJSRouting)) {
         Future(Redirect(controllers.routes.OtherReasonController.onPageLoadForUploadEvidence(mode)))
       } else {
         val nextPageIfNoUpload = navigation.nextPage(EvidencePage, mode)
@@ -153,9 +165,9 @@ class OtherReasonController @Inject()(whenDidBecomeUnablePage: WhenDidBecomeUnab
             upscanResponseModel => {
               val optErrorCode: Option[String] = request.session.get(SessionKeys.errorCodeFromUpscan)
               val optFailureFromUpscan: Option[String] = request.session.get(SessionKeys.failureMessageFromUpscan)
-              if(optErrorCode.isEmpty && optFailureFromUpscan.isEmpty) {
+              if (optErrorCode.isEmpty && optFailureFromUpscan.isEmpty) {
                 Ok(uploadFirstDocumentPage(upscanResponseModel, formProvider, nextPageIfNoUpload.url))
-              } else if(optErrorCode.isDefined && optFailureFromUpscan.isEmpty) {
+              } else if (optErrorCode.isDefined && optFailureFromUpscan.isEmpty) {
                 val localisedFailureReason = UpscanMessageHelper.getUploadFailureMessage(optErrorCode.get)
                 val formWithErrors = UploadDocumentForm.form.withError(FormError("file", localisedFailureReason))
                 BadRequest(uploadFirstDocumentPage(upscanResponseModel, formWithErrors, nextPageIfNoUpload.url))
@@ -184,16 +196,17 @@ class OtherReasonController @Inject()(whenDidBecomeUnablePage: WhenDidBecomeUnab
           upscanResponseModel => {
             val optErrorCode: Option[String] = request.session.get(SessionKeys.errorCodeFromUpscan)
             val optFailureFromUpscan: Option[String] = request.session.get(SessionKeys.failureMessageFromUpscan)
-            if(optErrorCode.isEmpty && optFailureFromUpscan.isEmpty) {
-              Ok(uploadAnotherDocumentPage(upscanResponseModel, formProvider))
-            } else if(optErrorCode.isDefined && optFailureFromUpscan.isEmpty) {
+            val fileListPage = navigation.nextPage(UploadAnotherDocumentPage, mode)
+            if (optErrorCode.isEmpty && optFailureFromUpscan.isEmpty) {
+              Ok(uploadAnotherDocumentPage(upscanResponseModel, formProvider, fileListPage.url))
+            } else if (optErrorCode.isDefined && optFailureFromUpscan.isEmpty) {
               val localisedFailureReason = UpscanMessageHelper.getUploadFailureMessage(optErrorCode.get)
               val formWithErrors = UploadDocumentForm.form.withError(FormError("file", localisedFailureReason))
-              BadRequest(uploadAnotherDocumentPage(upscanResponseModel, formWithErrors))
+              BadRequest(uploadAnotherDocumentPage(upscanResponseModel, formWithErrors, fileListPage.url))
                 .removingFromSession(SessionKeys.errorCodeFromUpscan)
             } else {
               val formWithErrors = UploadDocumentForm.form.withError(FormError("file", optFailureFromUpscan.get))
-              BadRequest(uploadAnotherDocumentPage(upscanResponseModel, formWithErrors))
+              BadRequest(uploadAnotherDocumentPage(upscanResponseModel, formWithErrors, fileListPage.url))
                 .removingFromSession(SessionKeys.failureMessageFromUpscan)
             }
           }
@@ -204,46 +217,42 @@ class OtherReasonController @Inject()(whenDidBecomeUnablePage: WhenDidBecomeUnab
   def onPageLoadForUploadComplete(mode: Mode): Action[AnyContent] = (authorise andThen dataRequired).async {
     implicit request => {
       val formProvider: Form[String] = FormProviderHelper.getSessionKeyAndAttemptToFillAnswerAsString(
-        YouHaveUploadedFilesForm.youHaveUploadedForm,
+        UploadListForm.youHaveUploadedForm,
         SessionKeys.nextFileUpload
       )
       val radioOptionsToRender: Seq[RadioItem] = RadioOptionHelper.yesNoRadioOptions(formProvider)
-      val postAction = controllers.routes.OtherReasonController.onSubmitForUploadComplete()
+      val postAction = controllers.routes.OtherReasonController.onSubmitForUploadComplete(mode)
       val journeyId = request.session.get(SessionKeys.journeyId).get
       upscanService.getAmountOfFilesUploadedForJourney(journeyId).map(
         filesUploaded => {
-            Ok(youHaveUploadedFilesPage(formProvider, radioOptionsToRender, postAction, filesUploaded))
-         }
+          Ok(uploadListPage(formProvider, radioOptionsToRender, postAction, filesUploaded))
+        }
       )
     }
   }
 
   def onSubmitForUploadComplete(mode: Mode): Action[AnyContent] = (authorise andThen dataRequired).async {
     implicit request => {
-      YouHaveUploadedFilesForm.youHaveUploadedForm.bindFromRequest.fold(
+      UploadListForm.youHaveUploadedForm.bindFromRequest.fold(
         formHasErrors => {
           val radioOptionsToRender: Seq[RadioItem] = RadioOptionHelper.yesNoRadioOptions(formHasErrors)
           val postAction = controllers.routes.OtherReasonController.onPageLoadForFirstFileUpload(mode)
           val journeyId = request.session.get(SessionKeys.journeyId).get
           upscanService.getAmountOfFilesUploadedForJourney(journeyId).map(
             filesUploaded => {
-              if(filesUploaded < 5) {
-                BadRequest(youHaveUploadedFilesPage(formHasErrors, radioOptionsToRender, postAction, filesUploaded))
+              if (filesUploaded < 5) {
+                BadRequest(uploadListPage(formHasErrors, radioOptionsToRender, postAction, filesUploaded))
               } else {
-                Redirect(controllers.routes.OtherReasonController.onPageLoadForUploadComplete()) //TODO routing to be updated
-                }
+                logger.debug("[OtherReasonController][onSubmitForUploadComplete] - Form has errors but 5 files have been uploaded - redirecting to next page")
+                Redirect(navigation.nextPage(FileListPage, NormalMode))
+              }
             }
           )
         },
         nextFileUpload => {
           logger.debug(
-            s"[YouHaveUploadedFilesController][onSubmitForNextFileUpload] - Adding '$nextFileUpload' to session under key: ${SessionKeys.nextFileUpload}")
-          Future(Redirect(
-            navigation.nextPage(
-              YouHaveUploadedFilesPage,
-              NormalMode,
-              Some(nextFileUpload)
-            ))
+            s"[OtherReasonController][onSubmitForUploadComplete] - Adding '$nextFileUpload' to session under key: ${SessionKeys.nextFileUpload}")
+          Future(Redirect(navigation.nextPage(FileListPage, mode, Some(nextFileUpload)))
             .addingToSession((SessionKeys.nextFileUpload, nextFileUpload)))
         }
       )
@@ -252,7 +261,7 @@ class OtherReasonController @Inject()(whenDidBecomeUnablePage: WhenDidBecomeUnab
 
   def removeFileUpload(mode: Mode): Action[AnyContent] = (authorise andThen dataRequired).async {
     implicit request => {
-      if(request.cookies.get("jsenabled").isDefined && !featureSwitching.isEnabled(NonJSRouting)) {
+      if (request.cookies.get("jsenabled").isDefined && !featureSwitching.isEnabled(NonJSRouting)) {
         Future(Redirect(controllers.routes.OtherReasonController.onPageLoadForUploadEvidence(mode)))
       } else {
         val journeyId = request.session.get(SessionKeys.journeyId).get
@@ -270,7 +279,7 @@ class OtherReasonController @Inject()(whenDidBecomeUnablePage: WhenDidBecomeUnab
                     logger.debug("[OtherReasonController][removeFileUpload] - No files left in journey - redirecting to first document upload page")
                     Redirect(controllers.routes.OtherReasonController.onPageLoadForFirstFileUpload(mode))
                   } else {
-                    Redirect(controllers.routes.OtherReasonController.onPageLoadForUploadComplete())
+                    Redirect(controllers.routes.OtherReasonController.onPageLoadForUploadComplete(mode))
                   }
                 }
               )
@@ -298,24 +307,25 @@ class OtherReasonController @Inject()(whenDidBecomeUnablePage: WhenDidBecomeUnab
         mode,
         isAddingAnotherDocument,
         {
-        (optFailureDetails, errorMessage) => {
-          if(errorMessage.isDefined) {
-            val failureReason = UpscanMessageHelper.getLocalisedFailureMessageForFailure(optFailureDetails.get.failureReason)
-            if(isAddingAnotherDocument) {
-              Future(Redirect(controllers.routes.OtherReasonController.onPageLoadForAnotherFileUpload(mode))
-                .addingToSession(SessionKeys.failureMessageFromUpscan -> failureReason)
-                .removingFromSession(SessionKeys.isAddingAnotherDocument, SessionKeys.fileReference))
+          (optFailureDetails, errorMessage) => {
+            if (errorMessage.isDefined) {
+              val failureReason = UpscanMessageHelper.getLocalisedFailureMessageForFailure(optFailureDetails.get.failureReason)
+              if (isAddingAnotherDocument) {
+                Future(Redirect(controllers.routes.OtherReasonController.onPageLoadForAnotherFileUpload(mode))
+                  .addingToSession(SessionKeys.failureMessageFromUpscan -> failureReason)
+                  .removingFromSession(SessionKeys.isAddingAnotherDocument, SessionKeys.fileReference))
+              } else {
+                Future(Redirect(controllers.routes.OtherReasonController.onPageLoadForFirstFileUpload(mode))
+                  .addingToSession(SessionKeys.failureMessageFromUpscan -> failureReason)
+                  .removingFromSession(SessionKeys.isAddingAnotherDocument, SessionKeys.fileReference))
+              }
             } else {
-              Future(Redirect(controllers.routes.OtherReasonController.onPageLoadForFirstFileUpload(mode))
-                .addingToSession(SessionKeys.failureMessageFromUpscan -> failureReason)
+              Future(Redirect(controllers.routes.OtherReasonController.onPageLoadForUploadComplete(mode))
                 .removingFromSession(SessionKeys.isAddingAnotherDocument, SessionKeys.fileReference))
             }
-          } else {
-            Future(Redirect(controllers.routes.OtherReasonController.onPageLoadForUploadComplete())
-              .removingFromSession(SessionKeys.isAddingAnotherDocument, SessionKeys.fileReference))
           }
-        }
-      })
+        })
     }
   }
 }
+
